refactor: Move code into search.* packages
[#OCD-3601]
