+++ chpl/chpl-api/src/main/java/gov/healthit/chpl/web/controller/SearchController.java
@@ -16,15 +16,15 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import gov.healthit.chpl.domain.search.ComplianceSearchFilter;
-import gov.healthit.chpl.domain.search.NonconformitySearchOptions;
-import gov.healthit.chpl.domain.search.OrderByOption;
-import gov.healthit.chpl.domain.search.SearchRequest;
-import gov.healthit.chpl.domain.search.SearchResponse;
-import gov.healthit.chpl.domain.search.SearchSetOperator;
 import gov.healthit.chpl.exception.InvalidArgumentsException;
 import gov.healthit.chpl.logging.Loggable;
-import gov.healthit.chpl.service.ListingSearchService;
+import gov.healthit.chpl.search.ListingSearchService;
+import gov.healthit.chpl.search.domain.ComplianceSearchFilter;
+import gov.healthit.chpl.search.domain.NonconformitySearchOptions;
+import gov.healthit.chpl.search.domain.OrderByOption;
+import gov.healthit.chpl.search.domain.SearchRequest;
+import gov.healthit.chpl.search.domain.SearchResponse;
+import gov.healthit.chpl.search.domain.SearchSetOperator;
 import gov.healthit.chpl.util.ErrorMessageUtil;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
