+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/search/CertifiedProductBasicSearchResultLegacy.java
@@ -5,6 +5,8 @@ import java.util.Set;
 
 import com.fasterxml.jackson.annotation.JsonView;
 
+import gov.healthit.chpl.search.domain.SearchViews;
+
 @Deprecated
 public class CertifiedProductBasicSearchResultLegacy extends CertifiedProductSearchResultLegacy {
     private static final long serialVersionUID = -2547390525592841038L;
