+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/search/CertifiedProductSearchResultLegacy.java
@@ -4,6 +4,7 @@ import java.io.Serializable;
 
 import com.fasterxml.jackson.annotation.JsonView;
 
+import gov.healthit.chpl.search.domain.SearchViews;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
