+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/search/domain/CertifiedProductBasicSearchResult.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.domain.search;
+package gov.healthit.chpl.search.domain;
 
 import java.util.HashSet;
 import java.util.Set;
