+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/search/domain/ComplianceSearchFilter.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.domain.search;
+package gov.healthit.chpl.search.domain;
 
 import java.io.Serializable;
 import java.util.HashSet;
