+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/search/domain/NonconformitySearchOptions.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.domain.search;
+package gov.healthit.chpl.search.domain;
 
 public enum NonconformitySearchOptions {
     NEVER_NONCONFORMITY, OPEN_NONCONFORMITY, CLOSED_NONCONFORMITY;
