+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/search/domain/SearchSetOperator.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.domain.search;
+package gov.healthit.chpl.search.domain;
 
 public enum SearchSetOperator {
     AND, OR;
